Refactor script into build and install portions
Now we have a build script (which builds the ANTLR tool and the C target
runtime on Darwin, and only the runtime on Linux) and a separate install
script (which installs tool and runtime on Darwin, and only the runtime
on Linux).

+++ antlr/build.sh
@@ -19,25 +19,31 @@ do
   patch -p0 < "${PATCH}"
 done
 
-echo "Building ANTLR..."
-OLD_CLASSPATH="$CLASSPATH"
-export CLASSPATH="/usr/local/junit/junit.jar:./lib/antlr-2.7.7.jar"
-cd "${ANTLR_VERSION}"
-ant -quiet generator
-cd codegen
-FILES=`ls *.java`
-cd ..
-for FILE in $FILES
-do
-  find src/org/antlr -name $FILE -delete
-done
-ant -quiet build
-export CLASSPATH="$OLD_CLASSPATH"
+# build ANTLR itself only on the local Darwin machine
+if [ `uname` = "Darwin" ]; then
+  echo "Building ANTLR..."
+  OLD_CLASSPATH="$CLASSPATH"
+  export CLASSPATH="/usr/local/junit/junit.jar:./lib/antlr-2.7.7.jar"
+  cd "${ANTLR_VERSION}"
+  ant -quiet generator
+  cd codegen
+  FILES=`ls *.java`
+  cd ..
+  for FILE in $FILES
+  do
+    find src/org/antlr -name $FILE -delete
+  done
+  ant -quiet build
+  cp build/antlr.jar lib/antlr-3.0.1-custom.jar
+  mv lib/antlr-3.0.1.jar lib/antlr-3.0.1.jar.disabled
+  export CLASSPATH="$OLD_CLASSPATH"
+  cd ..
+fi
 
 echo "Building C target runtime..."
-cd "runtime/C/dist/libantlr3c-${VERSION}"
+cd "${ANTLR_VERSION}/runtime/C/dist/libantlr3c-${VERSION}"
 ./configure --disable-shared
 make
 make check
 
-echo "Done"
+echo "Done: run ./install.sh to install"
